Moved to sync_providers
